Extract vue components to separate file and compile using vueify
The first step of the VueJS refactor is to extract all of the components
into separate `*.vue` files and import them as components to the main
Vue instance - or, as in the case of FriendsList - into the Conference
component.

Once the underlying component structure is settled on, we can start
working on actually refactoring the code as needed.

- [x] Extract components to separate files for vueify
- [ ] Establish an appropriate import style
- [ ] Make sure that each component works as it did previously