Tweak behaviour of add / delete items
To address an issue where the click event on the panel heading hijacked the delete button, apply the viewConference event on a (new) h3 only.

Bring in SweetAlert for delete confirmation, and use $remove to destroy items in the vue instance, rather than using underscore to filter properties.

Add an @ symbol as an input group addon on the friends list as this is prepended on display... and I kept ending up with @@ on my list.

Render the conference name on the conference view, so you know what conference you're looking at.