Prevent dashboard link appearing on introduction page
If I am logged in and introduce myself on another user's conference
page, I'll have a link to go back to the conference dashboard for that
user's conference.

Whilst the API shouldn't return any content for that conference for my
user, it's best not to link back to it in any case.