Adds new User Profile w/ public conference list and supporting changes
Adds a new user profile page that lists the user's upcoming and past conferences, allowing other users to browse conferences and easily introduce themselves with a button linking to the existing "Introduce yourself" link.

Adds a settings panel to allow the logged in user to turn on or off their public conference list.

Allows users to add a URL to a conference when creating, and changes conf badges to include a link to the conference page.

Adds tests for features mentioned above.